Fix up mg-desktop build post-sandbox-merge.
git-svn-id: http://svn.osgeo.org/mapguide/trunk@9268 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Desktop/MgDesktop/Services/Resource/ResourceContentCache.cpp
@@ -43,7 +43,7 @@ STRING MgdResourceContentCache::GetContentEntry(MgResourceIdentifier* resource)
 INT32 MgdResourceContentCache::GetCacheSize()
 {
     ACE_MT(ACE_GUARD_RETURN(ACE_Recursive_Thread_Mutex, ace_mon, m_mutex, 0));
-    return m_resourceContentCacheEntries.size();
+    return (INT32)m_resourceContentCacheEntries.size();
 }
 
 void MgdResourceContentCache::Clear()
